python/requirements.txt
kafka-python==2.0.2
mariadb==1.1.9
packaging==23.2
